+++ store/rootmultistore.go
@@ -183,10 +183,7 @@ func (rs *rootMultiStore) GetKVStore(key StoreKey) KVStore {
 	return rs.stores[key].(KVStore)
 }
 
-// Implements MultiStore.
-func (rs *rootMultiStore) GetKVStoreWithGas(meter sdk.GasMeter, key StoreKey) KVStore {
-	return NewGasKVStore(meter, rs.GetKVStore(key))
-}
+// Implements MultiStore
 
 // getStoreByName will first convert the original name to
 // a special key, before looking up the CommitStore.
@@ -267,6 +264,9 @@ func (rs *rootMultiStore) loadCommitStoreFromParams(id CommitID, params storePar
 		return
 	case sdk.StoreTypeDB:
 		panic("dbm.DB is not a CommitStore")
+	case sdk.StoreTypeTransient:
+		store = newTransientStore()
+		return
 	default:
 		panic(fmt.Sprintf("unrecognized store type %v", params.typ))
 	}
@@ -380,6 +380,10 @@ func commitStores(version int64, storeMap map[StoreKey]CommitStore) commitInfo {
 	for key, store := range storeMap {
 		// Commit
 		commitID := store.Commit()
+
+		if commitID.IsZero() {
+			continue
+		}
 
 		// Record CommitID
 		si := storeInfo{}